platform/pyproject.toml
[project]
name = "main"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "dagger-io",
    "pydantic>=2.9.1",
    "pyyaml>=6.0.2",
]

[tool.uv.sources]
dagger-io = { path = "sdk", editable = true }

[build-system]
requires = ["hatchling==1.25.0"]
build-backend = "hatchling.build"
